requirements.txt
tensorflow=2.7
tensorflow_text
numpy
django
django-cors-headers
djangorestframework
